[rescaleintensity]: (4) Moved funcitons from library/module to library/functions, created helper funciton in library/module, added pydantic annotations
